+++ Src/Base/AMReX_FileSystem.H
@@ -3,8 +3,17 @@
 
 #include <string>
 
+#ifdef _WIN32
+typedef unsigned short mode_t;
+#else
+#include <sys/types.h>  // for mode_t
+#endif
+
 namespace amrex {
 namespace FileSystem {
+
+bool
+CreateDirectories (std::string const& filename, mode_t mode, bool verbose = false);
 
 std::string
 CurrentPath ();
